+++ tests/test_user.py
@@ -1,4 +1,9 @@
 import pytest
+from easypost.constant import (
+    _FILTERS_KEY,
+    _TEST_FAILED_INTENTIONALLY_ERROR,
+    NO_MORE_PAGES_ERROR,
+)
 from easypost.models import (
     Brand,
     User,
@@ -76,3 +81,33 @@ def test_user_update_brand(prod_client):
     assert isinstance(brand, Brand)
     assert str.startswith(brand.id, "brd_")
     assert brand.color == color
+
+
+@pytest.mark.vcr()
+def test_user_all_children(prod_client, page_size):
+    children_data = prod_client.user.all_children(page_size=page_size)
+
+    children_array = children_data["children"]
+    assert len(children_array) <= page_size
+    assert all(isinstance(child, User) for child in children_array)
+
+    has_more = children_data["has_more"]
+    assert isinstance(has_more, bool)
+
+
+@pytest.mark.vcr()
+def test_user_children_get_next_page(prod_client, page_size):
+    try:
+        first_page = prod_client.user.all_children(page_size=page_size)
+        next_page = prod_client.user.get_next_page_of_children(children=first_page, page_size=page_size)
+
+        first_id_of_first_page = first_page["children"][0].id
+        first_id_of_second_page = next_page["children"][0].id
+
+        assert first_id_of_first_page != first_id_of_second_page
+
+        # Verify that the filters are being passed along for behind-the-scenes reference
+        assert first_page[_FILTERS_KEY] == next_page[_FILTERS_KEY]
+    except Exception as e:
+        if e.message != NO_MORE_PAGES_ERROR:
+            raise Exception(_TEST_FAILED_INTENTIONALLY_ERROR)