+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # CHANGELOG
 
+## Next release
+
+- Adds `all_children` function to the User service for retrieving paginated lists of children
+- Adds `get_next_page_of_children` function to User service to get next paginated list of children
+
 ## v9.0.1 (2023-12-20)
 
 - Corrects the return type of `regenerate_rates`